oracle-2/services/imaging-utils/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydicom==2.4.4
nibabel==5.2.1
numpy>=1.24.0
Pillow>=10.0.0
slowapi>=0.1.8
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

